rasa-student-bot/requirements.txt
# ============================================================
# requirements.txt
# ============================================================
rasa==3.6.20
rasa-sdk==3.6.2
pymongo[srv]==4.6.1
python-dotenv==1.0.0
